# MyRVM-Server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRVM API Documentation",
        "version": "1.0.0",
        "description": "API documentation for MyReverseVendingMachine Server",
        "contact": {
            "email": "admin@myrvm.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/devices/{id}/telemetry": {
            "post": {
                "tags": ["Edge Device"],
                "summary": "Send telemetry data",
                "operationId": "sendTelemetry",
                "security": [{"bearerAuth": []}],
                "parameters": [
                    { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["plastic_weight","aluminum_weight","total_items"],
                                "properties": {
                                    "plastic_weight": { "type": "number", "format": "float" },
                                    "aluminum_weight": { "type": "number", "format": "float" },
                                    "total_items": { "type": "integer" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "Telemetry recorded" },
                    "400": { "description": "Bad Request" },
                    "401": { "description": "Unauthorized - Invalid API Key" },
                    "404": { "description": "Device Not Found" },
                    "422": { "description": "Unprocessable Entity" },
                    "500": { "description": "Internal Server Error" }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}